+++ internal/tools/portinspect/addrinuse_other.go
@@ -1,0 +1,16 @@
+//go:build !windows
+
+package portinspect
+
+import (
+	"errors"
+	"syscall"
+)
+
+func isAddressInUse(err error) bool {
+	return errors.Is(err, syscall.EADDRINUSE)
+}
+
+func isPermissionDenied(err error) bool {
+	return errors.Is(err, syscall.EACCES) || errors.Is(err, syscall.EPERM)
+}